+++ thyrel-api/thyrel-api/Controllers/TestController.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using thyrel_api.Controllers.ModelsControllers;
 using thyrel_api.Models;
 using thyrel_api.Websocket;
